pyproject.toml
[project]
name = "idu-api"
version = "1.51.0"
description = "This is a Digital Territories Platform API in two versions to access and manipulate basic territories data."
authors = [
    {name = "Babayev Ruslan",email = "rus.babaef@yandex.ru"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "alembic[asyncio] (>=1.17.0,<2.0.0)",
    "geoalchemy2 (==0.15.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "geojson-pydantic (>=2.1.0,<3.0.0)",
    "fastapi-pagination (>=0.14.3,<0.15.0)",
    "sqlakeyset (>=2.0.1746777265,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "aioboto3 (>=15.4.0,<16.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tenacity (>=9.1.2,<10.0.0)",
    "prometheus-client (>=0.23.1,<0.24.0)",
    "otteroad (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "opentelemetry-exporter-otlp (>=1.39.0,<2.0.0)",
    "opentelemetry-exporter-prometheus (>=0.60b0,<0.61)",
    "opentelemetry-semantic-conventions (>=0.60b0,<0.61)",
    "opentelemetry-instrumentation-logging (>=0.60b0,<0.61)",
]

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pylint (>=4.0.2,<5.0.0)",
    "pytest[asyncio,coverage] (>=8.4.2,<9.0.0)",
    "tqdm[asyncio] (>=4.67.1,<5.0.0)",
    "aioresponses (>=0.7.8,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
launch_urban_api = "idu_api.urban_api.__main__:main"
# launch_city_api = "idu_api.city_api.__main__:main" # TODO: add launch option here


[tool.black]
line-length = 120
target-version = ['py311']

[tool.pylint.format]
max-line-length = 120
expected-line-ending-format = "LF"
max-locals = 20
max-attributes = 10
max-args = 9
max-positional-arguments = 12
extension-pkg-allow-list = ["networkit"]
ignored-modules = ["alembic.context"]
disable = [
    "duplicate-code",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
split_on_trailing_comma = true
profile = "black"
